Moved helperfunctions to it pruchase-helpers.ts and made purchase-action use server
